pyproject.toml
[project]
name = "pitstoptimingparsing"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-rest-framework>=0.1.0",
    "djangorestframework-stubs>=3.16.8",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.29.0",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "pendulum>=3.2.0",
    "psycopg2>=2.9.11",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "respx>=0.22.0",
    "rich>=14.3.3",
    "ruff>=0.15.0",
    "typer>=0.24.0",
]

[tool.ruff]
# Эквивалент Black + isort + flake8
line-length = 88
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
# Включаем правила: pyflakes, pycodestyle, isort, pydocstyle, etc.
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "D",   # pydocstyle
  "UP",  # pyupgrade
  "C4",  # flake8-comprehensions
]
ignore = [
  "D100", "D104", "D105", "D107",  # Missing docstring in public module/function/etc.
  "E501",  # line-too-long (ruff сам форматирует)
]

[tool.ruff.lint.isort]
known-first-party = ["core", "api", "parser", "config"]

[tool.ruff.format]
# Форматирование в стиле Black
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "google"
